+++ python/mxnet/contrib/autograd.py
@@ -8,7 +8,6 @@ import functools
 from ..base import _LIB, check_call, string_types
 from ..base import mx_uint, NDArrayHandle, c_array
 # pylint: disable= unused-import
-from ..ndarray_utils import *
 from ..ndarray import NDArray, zeros_like
 from ..symbol import _GRAD_REQ_MAP
